pyproject.toml
[tool.poetry]
authors = ["Harvard Library Innovation Lab"]
description = "A simple REST API for witnessing the web using the Scoop web archiving capture engine."
license = "MIT"
name = "scoop-witness-api"
packages = [{include = "scoop_witness_api"}]
readme = "README.md"
version = "0.0.1"

[tool.poetry.dependencies]
Flask = "^2.3.2"
bcrypt = "^4.0.1"
black = "^23.7.0"
cryptography = "^41.0.2"
gunicorn = "^20.1.0"
peewee = "^3.16.2"
pymysql = "^1.1.0"
python = "^3.9"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
validators = "^0.20.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.3.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 100
